Almost completed main ZWaveLib rework. Some concerns about the MultiInstance class, this should be tested. Also MIG ZWave interface "Toggle" command is to be fixed.
