--- MigFiles/SupportLibraries/ZWaveLib/Devices/ProductHandlers/Zwave.Me/ZME_06436MotorControl.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ZWaveLib.Devices.ProductHandlers.ZwaveME
-{
-    public class ZME_06436MotorControl : ProductHandlers.Generic.Switch
-    {
-
-        public override bool CanHandleProduct(ManufacturerSpecific productspecs)
-        {
-            return (productspecs.ManufacturerId == "0115" && productspecs.TypeId == "1000" && productspecs.ProductId == "0003");
-        }
-
-    }
-}
