+++ MigFiles/SupportLibraries/ZWaveLib/Handlers/SwitchBinary.cs
@@ -21,7 +21,6 @@
  */
 
 using System;
-using ZWaveLib.Devices;
 
 namespace ZWaveLib.Handlers
 {
