+++ MigFiles/SupportLibraries/ZWaveLib/ZWaveMessage.cs
@@ -24,7 +24,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using ZWaveLib.Devices;
 
 namespace ZWaveLib
 {
